decision-rag/backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
apscheduler==3.10.4
azure-identity==1.15.0
azure-storage-blob==12.19.0
pyarrow==15.0
black==24.8.0
pgvector==0.4.2
psycopg2-binary==2.9.7
certifi==2025.11.12
click==8.1.8 # On versions >=8.2.0 there is a bug causing issues with Typer (https://github.com/ai-dynamo/dynamo/issues/1039)
coverage==7.13.0
elasticsearch==8.16.0
fastapi==0.115.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
markdown-it-py==4.0.0
docling==2.80.0
mdurl==0.1.2
mypy==1.11.2
mypy_extensions==1.1.0
openai==1.59.5
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.19.2
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dateutil==2.9.0
python-dotenv==1.0.1
rich==13.9.2
ruff==0.6.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tenacity==9.0.0
tiktoken==0.8.0
typer==0.15.3
typing_extensions==4.15.0
uvicorn[standard]==0.32.0
sentry-sdk==2.53.0
